requirements.txt
streamlit
yfinance
pandas
plotly
google-genai
beautifulsoup4
requests